change graphql struct
