+++ app/graphql/types/mutation_type.rb
@@ -2,6 +2,6 @@
 
 module Types
   class MutationType < Types::BaseObject
-    field :save_objective, mutation: Mutations::Questions::Objective::Save
+    field :save_objective, mutation: Mutations::Questions::SaveObjective
   end
 end
